Fix: Mesasges Component refactor
